+++ melodyd/coreaudio_other.go
@@ -1,0 +1,7 @@
+//go:build !darwin
+
+package main
+
+func listCoreAudioOutputDevices() ([]coreAudioOutputDevice, error) {
+	return nil, nil
+}